requirements.txt
asgiref==3.8.1
boto3==1.35.90
botocore==1.35.90
click==8.1.8
crispy-bootstrap4==2024.10
dj-database-url==0.5.0
#Django==5.1.3
Django==4.2.10  # Replace with the latest patch version
django-allauth==0.50.0
django-countries==7.6.1
django-crispy-forms==2.3
django-storages==1.14.4
git-filter-repo==2.47.0
gunicorn==23.0.0
jmespath==1.0.1
oauthlib==3.2.2
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.10.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests-oauthlib==2.0.0
s3transfer==0.10.4
setuptools==75.6.0
sqlparse==0.5.2
stripe==11.4.1
wheel==0.45.1
